Generic charter datasets classes.

+++ collate_xml_to_img.sh
@@ -16,9 +16,14 @@ fi
 test -d $1 || mkdir $1
 echo $1
 for xml in $(find ./exports -name "*.xml" | grep -Ev '(mets.xml|metadata.xml)') ; do
-	echo $xml
        	img_file=$(grep imageFilename $xml | sed 's/^.*imageFilename="\([^\"]\+\)".*$/\1/' ) ;
 	echo $img_file
-       	cp Digitalisate_JPG/$img_file $xml $1
+	if [[ -f "Digitalisate_JPG/$img_file" ]]; then
+		cp Digitalisate_JPG/$img_file $xml $1
+	else
+		found=$(find . -name $img_file -type f | head -1)
+		test -n "$found" && cp $found $xml $1
+	fi
+
 done
